機材ページでのstate管理をhooksからcontextに変更、shipページの構成変更
