+++ src/components/utils/Modal.tsx
@@ -1,14 +1,18 @@
-import * as React from 'react';
+import React, {useContext} from 'react';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import Modal from '@mui/material/Modal';
+
+import {ShipContext} from "components/providers/ShipContextProvider"
 
 
 export const BasicModal = () => {
   const [open, setOpen] = React.useState(false);
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
+
+  const {selectShip} = useContext(ShipContext)
 
   const style = {
     position: 'absolute' as 'absolute',
@@ -34,7 +38,7 @@ export const BasicModal = () => {
       >
         <Box sx={style}>
           <Typography id="modal-modal-title" variant="h6" component="h2">
-            Text in a modal
+            {selectShip}
           </Typography>
           <Typography id="modal-modal-description" sx={{ mt: 2 }}>
             Duis mollis, est non commodo luctus, nisi erat porttitor ligula.
